requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
async-timeout==4.0.3
asyncclick==8.1.7.2
attrs==23.2.0
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
Flask==3.0.3
frozenlist==1.4.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
python-kasa==0.6.2.1
requests==2.32.2
sniffio==1.3.1
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
yarl==1.9.4
